requirements.txt
alembic==1.8.1
APScheduler==3.9.1
click==8.1.3
colorama==0.4.5
Flask==2.1.3
Flask-APScheduler==1.12.4
Flask-Login==0.6.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.1
MarkupSafe==2.1.1
PyMySQL==1.0.2
pyserial==3.5
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
SQLAlchemy==1.4.39
tzdata==2022.1
tzlocal==4.2
Werkzeug==2.1.2
WTForms==3.0.1
xlwt==1.3.0
zipp==3.8.1
